change to uint256

+++ contracts/NZCovidBadge.sol
@@ -17,11 +17,10 @@ contract NZCOVIDBadge is ERC721, Verifier, EllipticCurve {
     }
 
     function mint(
-            uint[2] memory a,
-            uint[2][2] memory b,
-            uint[2] memory c,
-            uint[3] memory input, bytes32 toBeSignedHash, uint256[2] memory rs) public payable {
-
+            uint256[2] memory a,
+            uint256[2][2] memory b,
+            uint256[2] memory c,
+            uint256[3] memory input, bytes32 toBeSignedHash, uint256[2] memory rs) public payable {
         uint exampleX = 0xCD147E5C6B02A75D95BDB82E8B80C3E8EE9CAA685F3EE5CC862D4EC4F97CEFAD;
         uint exampleY = 0x22FE5253A16E5BE4D1621E7F18EAC995C57F82917F1A9150842383F0B4A4DD3D;
